finish backend-2

+++ job/views.py
@@ -1,11 +1,14 @@
 from django.shortcuts import render
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view,permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 from django.db.models import Avg, Min, Max, Count 
 from rest_framework.pagination import PageNumberPagination
-from .serializers import JobSerializer
-from .models import Job
+from rest_framework.permissions import IsAuthenticated
+from django.utils import timezone
+
+from .serializers import JobSerializer, CandidatesAppliedSerializer
+from .models import Job,CandidatesApplied
 
 from django.shortcuts import get_object_or_404
 from .filters import JobsFilter
@@ -44,7 +47,9 @@ def getJob(request, pk):
     return Response(serializer.data)
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
 
     job = Job.objects.create(**data)
@@ -53,8 +58,12 @@ def newJob(request):
     return Response(serializer.data)
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({ 'message': 'You can not update this job' }, status=status.HTTP_403_FORBIDDEN)
 
     job.title = request.data['title']
     job.description = request.data['description']
@@ -74,8 +83,13 @@ def updateJob(request, pk):
     return Response(serializer.data)
 
 @api_view(['DELETE'])
-def deleteJob(reqeust, pk):
+@permission_classes([IsAuthenticated])
+def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({ 'message': 'You can not delete this job' }, status=status.HTTP_403_FORBIDDEN)
+
 
     job.delete()
 
@@ -98,3 +112,77 @@ def getTopicStats(request, topic):
     )
 
     return Response(stats)
+
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+def applyToJob(request, pk):
+    user = request.user
+    job = get_object_or_404(Job, id=pk)
+
+    if user.userprofile.resume == '':
+        return Response({'error': 'Please upload your resume first'}, status=status.HTTP_400_BAD_REQUEST)
+    
+    if job.lastDate < timezone.now():
+        return Response({'error': 'You can not apply to this job. Date is over'}, status=status.HTTP_400_BAD_REQUEST)
+    
+    alreadyApplied = job.candidatesapplied_set.filter(user=user).exists()
+
+    if alreadyApplied:
+        return Response({'error': 'You have already apply to this job.'}, status=status.HTTP_400_BAD_REQUEST)
+    
+    jobApplied = CandidatesApplied.objects.create(
+        job = job,
+        user = user,
+        resume = user.userprofile.resume
+    )
+
+    return Response({
+        'applied': True,
+        'job_id': jobApplied.id
+    },
+    status=status.HTTP_200_OK
+    )
+
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def getCurrentUserAppliedJobs(request):
+    args = {'user_id': request.user.id}
+
+    jobs = CandidatesApplied.objects.filter(**args)
+
+    serializer = CandidatesAppliedSerializer(jobs, many=True)
+
+    return Response(serializer.data)
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def isApplied(request, pk):
+
+    user = request.user
+    job = get_object_or_404(Job, id=pk)
+
+    applied = job.candidatesapplied_set.filter(user=user).exists()
+
+    return Response(applied)
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def getCurrentUserJobs(request):
+    args = {'user_id': request.user.id}
+    jobs = Job.objects.filter(**args)
+    serializer = JobSerializer(jobs, many=True)
+    return Response(serializer.data)
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def getCandidatesApplied(request):
+    user = request.user
+    job = get_object_or_404(Job, id=pk)
+
+    if job.user != user:
+        return Response({ 'message': 'You can not access this job' }, status=status.HTTP_403_FORBIDDEN)
+    candidates = job.candidatesapplied_set.all()
+
+    serializer = CandidatesAppliedSerializer(candidates, many=True)
+    return Response(serializer.data)
